<fix>: move user-project entity
